pyproject.toml
[tool.poetry]
name = "oepul-chat"
version = "0.1.0"
description = ""
authors = ["vali101 <valentin.peter@hotmail.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
llama-index = "^0.9.39"
langchain = "^0.1.4"
langchain-openai = "^0.0.5"
pdfplumber = "^0.10.3"
selenium = "^4.17.2"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
